Symlink files and folders by convention
Files or folders ending with '.symlink' and '.<os>-symlink' will be
symlinked to $HOME.

If a file or folder needs to be symlinked to from a different location
depending on the OS, a config file should be added as
'<file-or-folder-name>.symlinks'. The content of the file should have
the following structure:

    <os1>: path/to/file1
    <os2>: path/to/file2
    ...

See 'vim/.vim.symlinks' for an example.